ConfirmationDialogs added to remove/delete functions

+++ src/main/java/ch/jtaf/ui/component/GridBuilder.java
@@ -4,6 +4,7 @@ import ch.jtaf.ui.dialog.EditDialog;
 import ch.jtaf.ui.function.Callback;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.button.ButtonVariant;
+import com.vaadin.flow.component.confirmdialog.ConfirmDialog;
 import com.vaadin.flow.component.grid.ColumnTextAlign;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.notification.Notification;
@@ -37,18 +38,28 @@ public class GridBuilder {
         grid.addComponentColumn(record -> {
             Button delete = new Button(insteadOfDeleteTitle != null ? insteadOfDeleteTitle : grid.getTranslation("Delete"));
             delete.addThemeVariants(ButtonVariant.LUMO_ERROR);
-            delete.addClickListener(event -> getBean(TransactionTemplate.class).executeWithoutResult(transactionStatus -> {
-                try {
-                    if (insteadOfDelete != null) {
-                        insteadOfDelete.accept(record);
-                    } else {
-                        getBean(DSLContext.class).attach(record);
-                        record.delete();
-                    }
-                } catch (DataAccessException e) {
-                    Notification.show(e.getMessage());
-                }
-            }));
+            delete.addClickListener(event -> {
+                ConfirmDialog confirmDialog = new ConfirmDialog(grid.getTranslation("Confirm"),
+                    grid.getTranslation("Are.you.sure"),
+                    grid.getTranslation("Delete"), e ->
+                    getBean(TransactionTemplate.class).executeWithoutResult(transactionStatus -> {
+                        try {
+                            if (insteadOfDelete != null) {
+                                insteadOfDelete.accept(record);
+                            } else {
+                                getBean(DSLContext.class).attach(record);
+                                record.delete();
+                            }
+                        } catch (DataAccessException ex) {
+                            Notification.show(ex.getMessage());
+                        }
+                    }),
+                    grid.getTranslation("Cancel"), e -> {
+                });
+                confirmDialog.setConfirmButtonTheme("error primary");
+                confirmDialog.open();
+
+            });
             HorizontalLayout horizontalLayout = new HorizontalLayout(delete);
             horizontalLayout.setJustifyContentMode(FlexComponent.JustifyContentMode.END);
             return horizontalLayout;
